Separate causal constant for easier merging
