+++ corebehrt/constants/causal/paths.py
@@ -1,0 +1,21 @@
+## Causal ##
+ENCODE_CFG = "encode_config.yaml"
+SIMULATE_CFG = "simulate_config.yaml"
+CALIBRATE_CFG = "calibrate_config.yaml"
+TRAIN_MLP_CFG = "train_mlp_config.yaml"
+ESTIMATE_CFG = "estimate_config.yaml"
+ENCODINGS_FILE = "encodings.pt"
+
+# Files
+PREDICTIONS_FILE = "predictions_and_targets.csv"
+CALIBRATED_PREDICTIONS_FILE = "predictions_and_targets_calibrated.csv"
+
+TIMESTAMP_OUTCOME_FILE = "outcomes_with_timestamps.csv"
+SIMULATION_RESULTS_FILE = "probas_and_outcomes.csv"
+
+EXPERIMENT_DATA_FILE = "experiment_data.parquet"
+EXPERIMENT_STATS_FILE = "experiment_stats.csv"
+ESTIMATE_RESULTS_FILE = "estimate_results.csv"
+
+SKS_DUMP_DIR = "corebehrt/main_causal/helper/data/sks_dumps"
+SKS_TREES_DIR = "corebehrt/main_causal/helper/data/sks_trees"
