+++ corebehrt/main_causal/simulate.py
@@ -1,7 +1,10 @@
 import logging
 from os.path import join
 
-from corebehrt.constants.causal import SIMULATION_RESULTS_FILE, TIMESTAMP_OUTCOME_FILE
+from corebehrt.constants.causal.paths import (
+    SIMULATION_RESULTS_FILE,
+    TIMESTAMP_OUTCOME_FILE,
+)
 from corebehrt.constants.data import ABSPOS_COL, TIMESTAMP_COL
 from corebehrt.functional.causal.load import (
     load_encodings_and_pids_from_encoded_dir,
